backend_fastapi/pyproject.toml
[tool.poetry]
name = "backend-fastapi"
version = "0.1.0"
description = ""
authors = ["rizal.mujahiddan <rizal.mujahiddan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
transformers = "^4.47.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
